backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases
pydantic
python-jose
passlib[bcrypt]
python-multipart